+++ carts/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 
-from carts.views import CartToOrderView, CartView
+from carts.views import CartView
 
 urlpatterns = [
-    path('', CartView.as_view()),
-    path('/neworder', CartToOrderView.as_view())
+    path('', CartView.as_view())
 ]
