+++ carts/views.py
@@ -6,8 +6,6 @@ from django.views    import View
 from carts.models    import Cart
 from users.utils     import signin_decorator
 from products.models import Product
-from orders.models   import Order, OrderProduct
-
 
 class CartView(View):
     @signin_decorator
@@ -16,7 +14,7 @@ class CartView(View):
             data     = json.loads(request.body)
             quantity = data['quantity']
             product  = Product.objects.get(id=data['product_id'])
-
+            
             cart, created = Cart.objects.get_or_create(
                 defaults  = {'quantity' : quantity},
                 product   = product,
@@ -76,32 +74,3 @@ class CartView(View):
 
         except Cart.DoesNotExist:
             return JsonResponse({'message':'JSONDecodeError'}, status=404)
-
-class CartToOrderView(View):
-    @signin_decorator
-    def post(self,request):
-        try:
-            data  = json.loads(request.body)
-            carts = Cart.objects.filter(id__in=data['cart_ids'])
-            
-            order = Order.objects.create(
-                order_status_id = 1,
-                user_id         = request.user.id
-            )
-
-            for cart in carts:
-                OrderProduct.objects.create(
-                    order_id                = order.id,
-                    product_id              = cart.product_id,
-                    quantity                = cart.quantity,
-                    order_product_status_id = 1
-                )
-
-            Cart.objects.filter(user=request.user, id__in=data['cart_ids']).delete()
-            return JsonResponse({"message" : "NEW_ORDER_CREATED"}, status=201)
-
-        except json.JSONDecodeError:
-            return JsonResponse({'message':'JSONDecodeError'}, status=404)
-
-        except KeyError:
-            return JsonResponse({"message" : "KEY_ERROR"}, status=400)
