+++ orders/urls.py
@@ -1,4 +1,7 @@
-from django.urls import path
+from django.urls  import path
+
+from orders.views import OrderView
 
 urlpatterns = [
+    path('/neworder', OrderView.as_view())
 ]
